fix util
